Works with 5.2.5.0
1. Lots of bugfixes.
2. Public functions and classes are commented since it seems number of users > 1.
3. Lots of things are more robust.
4. Data is now returned either as a numpy structured array or as a namedtuple. No more dicts. If you want a dict, just use asdict()
5. You can now access news, via NewsConn. Real time news was always available but now we have historical news, news counts, full news stories etc.

Finally got the time to rewrite the changes requested in some pull requests that weren't doing things the way I would want. Code has the same or better functionality but is idiomatic Python, simple and doesn't crash.

service.py is much more robust. Instead of waiting 5 seconds after launching IQFeed.exe, it checks to see if it can connect to the relevant sockets. Boilerplate file called launch.py created. Use this as a starting point for a cron job that stars IQFEed and monitors it. Created a new file launch_iqfeed.py which shows how FeedService should be used and has copious comments on it's use in an actual trading environment.

__init__.py imports only the classes you need.

TODO:
- Names used in field_readers.py need to be changed to refer to what is being read. Right now the file is a mess.
- conn.py and listeners.py should probably be divvied up in files in a better way. conn.py is too long. Should each connection class and the appropriate listeners be in a separate file?
- Market depth data.
- Need to parse condition codes etc.